cambios nuevos

+++ src/main/java/ar/edu/iua/iw3/cuentas/UserRepository.java
@@ -7,9 +7,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import ar.edu.iua.iw3.modelo.Producto;
+import ar.edu.iua.iw3.negocio.excepciones.EncontradoException;
+import ar.edu.iua.iw3.negocio.excepciones.NegocioException;
+import ar.edu.iua.iw3.negocio.excepciones.NoEncontradoException;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
  
-	public Optional<User> findFirstByUsernameOrEmail(String username, String email);
+public Optional<User> findFirstByUsernameOrEmail(String username, String email);
+				
 }
